backend/requirements.txt
vision-agents[gemini,getstream]==0.3.8
vision-agents-plugins-cartesia==0.3.8
python-dotenv==1.2.1
roboflow==1.2.14
supervision==0.27.0.post1
opencv-python==4.10.0.84
fastapi>=0.128.0
uvicorn>=0.38.0
numpy>=1.26.0
aiortc>=1.9.0
av>=12.0.0
ultralytics>=8.3.0
scipy>=1.11.0
tensorflow==2.20.0
tensorflow_hub==0.16.1
